modularize miltiplier tests
